move example app to com.topsort.example
