+++ TopsortAnalytics/src/main/java/com/topsort/analytics/banners/BannerView.kt
@@ -1,19 +1,23 @@
 package com.topsort.analytics.banners
 
 import android.content.Context
+import android.util.AttributeSet
 import android.widget.ImageView
 import coil.load
 import com.topsort.analytics.Analytics
 import com.topsort.analytics.model.Placement
 import com.topsort.analytics.model.auctions.EntityType
 
+
 class BannerView(
     context: Context,
-    config: BannerConfig,
-    path: String,
-    location: String,
-    clickCallback: (String, EntityType) -> Unit
-) : ImageView(context) {
+    attrs: AttributeSet
+) : ImageView(context, attrs) {
+    lateinit var config: BannerConfig
+    lateinit var location: String
+    lateinit var path: String
+    lateinit var onClick: (String, EntityType) -> Unit
+
     init {
         val result = runBannerAuction(config)
         if (result != null) {
@@ -29,7 +33,7 @@ class BannerView(
                     resolvedBidId = result.resolvedBidId,
                     placement = Placement(path = path, location = location)
                 )
-                clickCallback(result.id, result.type)
+                onClick(result.id, result.type)
             }
         }
     }
